changes for edit patient functionality

+++ healthCare-Frontend/src/components/patients-list/PatientsList.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 import { useContext, useState } from "react";
 import { Alert, Table } from "react-bootstrap";
-import { AiOutlineDelete, AiOutlineEye, AiOutlineEdit } from "react-icons/ai";
+import { AiOutlineDelete, AiOutlineEdit, AiOutlineEye } from "react-icons/ai";
 import { useNavigate } from "react-router-dom";
 import DataContext from "../../context/DataContext";
 import { BASE_API_URL } from "../../utils/constants";
@@ -11,51 +11,51 @@ const PatientsList = ({ filteredData, setFilteredData }) => {
   const navigate = useNavigate();
   const [successMsg, setSuccessMsg] = useState("");
   const [errorMsg, setErrorMsg] = useState("");
-  const { setData, setIsReadOnly, setStep, setIsEdit, isEdit } =
+  const { setData, setIsReadOnly, setStep, setIsEdit, setPatientId } =
     useContext(DataContext);
 
-  const handleViewDetailsClick = async (patient_id) => {
+  const handleViewDetailsClick = async (patient_id, selectedOption) => {
     try {
+      setPatientId(patient_id);
       const { data: details } = await axios.get(
         `${BASE_API_URL}/v1/patient/${patient_id}`
       );
       // const { data: details } = await axios.get(`/patient.json`);
       setData(details);
       setStep(1);
-      setIsReadOnly(true);
+      if (selectedOption === "view") {
+        setIsEdit(false);
+        setIsReadOnly(true);
+      } else if (selectedOption === "edit") {
+        setIsReadOnly(false);
+        setIsEdit(true);
+      }
       navigate("/step-one");
-    } catch (error) {}
-  };
-
-  const handleDeleteClick = async (patient_id) => {
-    try {
-      await axios.delete(`${BASE_API_URL}/v1/patient/${patient_id}`);
-      setSuccessMsg("Data is successfully deleted.");
-      setFilteredData(
-        filteredData.filter((patient) => patient.patient_id !== patient_id)
-      );
-      setErrorMsg("");
-      setTimeout(() => {
-        setSuccessMsg("");
-      }, 5000);
     } catch (error) {
-      setSuccessMsg("");
-      setErrorMsg("Error while deleting data. Please try again.");
+      console.log("error while viewing details. Try again later.");
     }
   };
 
-  const handleEditClick = async (patient_id) => {
-    try {
-      const { data: details } = await axios.get(
-        `${BASE_API_URL}/v1/patient/${patient_id}`
-      );
-      // const { data: details } = await axios.get(`/patient.json`);
-      setData(details);
-      setStep(1);
-      setIsReadOnly(false);
-      setIsEdit(true);
-      navigate("/step-one");
-    } catch (error) {}
+  const handleDeleteClick = async (patient_id, motherName) => {
+    const response = window.confirm(
+      `Are you sure you want to delete patient data for mother name ${motherName}?`
+    );
+    if (response) {
+      try {
+        await axios.delete(`${BASE_API_URL}/v1/patient/${patient_id}`);
+        setSuccessMsg("Data is successfully deleted.");
+        setFilteredData(
+          filteredData.filter((patient) => patient.patient_id !== patient_id)
+        );
+        setErrorMsg("");
+        setTimeout(() => {
+          setSuccessMsg("");
+        }, 5000);
+      } catch (error) {
+        setSuccessMsg("");
+        setErrorMsg("Error while deleting data. Please try again.");
+      }
+    }
   };
 
   return (
@@ -91,21 +91,21 @@ const PatientsList = ({ filteredData, setFilteredData }) => {
                   <AiOutlineEye
                     size={30}
                     className="icon"
-                    onClick={() => handleViewDetailsClick(patient_id)}
+                    onClick={() => handleViewDetailsClick(patient_id, "view")}
                   />
                 </td>
                 <td className="icon">
                   <AiOutlineEdit
                     size={30}
                     className="icon"
-                    onClick={() => handleEditClick(patient_id)}
+                    onClick={() => handleViewDetailsClick(patient_id, "edit")}
                   />
                 </td>
                 <td className="icon">
                   <AiOutlineDelete
                     size={30}
                     className="icon"
-                    onClick={() => handleDeleteClick(patient_id)}
+                    onClick={() => handleDeleteClick(patient_id, motherName)}
                   />
                 </td>
               </tr>
